Fix Commit: Note: there might be nothing wrong with the bpt, but the "Mystring" went wrong, I have fixed this problem, now with the map serving as bpt, the first two tasks passed Date:2024.5.20 21:07

+++ src/include/vector.hpp
@@ -327,7 +327,7 @@ public:
    * clears the contents
    */
   void clear() {
-    for (int i = 0; i < current_size; i++) {
+    for (int i = 0; i < (int)current_size; i++) {
       std::destroy_at(begin_ptr + i);
     }
     alloc.deallocate(begin_ptr, max_size);
